login-server-requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
Flask_MySQLdb==2.0.0
Pillow==9.4.0
Pillow==10.1.0
